Making SigningKey and EncryptionKeys compatible with processors
