+++ lib/src/commonTest/kotlin/co/touchlab/kjwt/JweEncodeTest.kt
@@ -7,8 +7,7 @@ import co.touchlab.kjwt.cryptography.ext.encryptWith
 import co.touchlab.kjwt.ext.subjectOrNull
 import co.touchlab.kjwt.model.algorithm.EncryptionAlgorithm
 import co.touchlab.kjwt.model.algorithm.EncryptionContentAlgorithm
-import co.touchlab.kjwt.cryptography.registry.EncryptionKey
-import dev.whyoleg.cryptography.algorithms.RSA
+import co.touchlab.kjwt.cryptography.processors.EncryptionKey
 import io.kotest.core.spec.style.FunSpec
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
