+++ lib/src/commonTest/kotlin/co/touchlab/kjwt/JwtKeyRegistryTest.kt
@@ -11,8 +11,8 @@ import co.touchlab.kjwt.ext.subjectOrNull
 import co.touchlab.kjwt.model.algorithm.EncryptionAlgorithm
 import co.touchlab.kjwt.model.algorithm.EncryptionContentAlgorithm
 import co.touchlab.kjwt.model.algorithm.SigningAlgorithm
-import co.touchlab.kjwt.cryptography.registry.EncryptionKey
-import co.touchlab.kjwt.cryptography.registry.SigningKey
+import co.touchlab.kjwt.cryptography.processors.EncryptionKey
+import co.touchlab.kjwt.cryptography.processors.SigningKey
 import co.touchlab.kjwt.model.registry.DefaultJwtKeyRegistry
 import io.kotest.core.spec.style.FunSpec
 import kotlin.test.assertEquals
